Removed nodemon from Node server
Had sporadic issues with the server not starting (because it says "not enough memory"), crashing on start/restart, etc.  The issues were so bad that it would lock the port on my computer and I couldn't shutdown/restart the servers.  The only way I could fix it was to restart my entire PC.  I couldn't narrow down the exact problem, but nodemon seems to be a factor in it?