Added dropdown arrow to instructor-reports component
I have added the drop down arrow to the instructor-reports component and attached the click event listener to the image so now you press the button to expand the student.
However, I am having difficulty working out how to have the image rotated 180 on the horizontal axis to have it become a collapse button after you press it.