My work for 10/4/2023
Mostly implemented the logic to replace the temporary data with actual data for the viewing time logs for all users.
The only issue is that it will only do so for those that have created a time card for their project and not for all users, regardless of if they have made a time table or are part of a project/group.  It also only displays the raw value for the data, like 20757 and  11163, not the converted value into hours, minutes, and seconds.  Every 1000 value is a second for the raw data from what I can gather.

I also replaced all API call to http://localhost:8080/api/Add-Courses with http://localhost:8080/api/Courses because they do the same thing.  However from what I see, component is not used?  Have to talk to group  if we should remove it.