Finished replacing temporary data with actual data I think
The temporary data is now replaced with the actual data from the database.  It will display all students that are registered for the course and will grab all the projects and the total times for those projects they are assigned to.  Even if they are not assigned to a project, they will be displayed.

There however might be an error in logic when joining to the table "Project_Users" in the SQL statement.  See the file TimeCardControllers.js for the information.