Braxton's work for 10/6/2023
Fixed bug with the "GetTotalTimeForAllUsersInCourse" now renamed to "GetReportsData" controller so it now mostly displays the projects correctly.  However, there is still an issue with the sql statment because take the below example.

A user is part of a group and has made some timecards, but then they are voted out or they leave.

The above situation would cause it so that the name of the project would be null, thus not displaying the project to the user, even though they have time in it.

I also started to implement the new page that will displays all the time logs for a user working on a project.