Implemented the detailed report
I have finished implementing the view-report component to allow you to view the detailed report for any user for any project they are a part of.

it displays the time in, time out, and description columns from the database so the user has a broken down view of what was done in the project for that user.