Started to implment the delete account functionality
Currently the route is set up using the http delete request to receive the request to delete the user.  However, I seem to be having problems passing the userID to the route inside the body of the request.  the body is empty when I debug it, might move to a post request as the deletedCourse route uses?