Started to implement manual time card entries
Currently the user can switch between auto and manual mode for entering a time card.  The logic is also there to make sure that the submit buttons are disabled until the form is valid for manual mode and when the form is valid and the timer is running for auto mode.

Currently, there is no implementation to determine how much time is between the two dates and checks in place to make sure that the the end date is latter than the start date for manual mode.

It shows that there is a lot of changes in the two files, but that is because I auto formatted the documents.