Manual entry for timecards now mostly implemented
Timecards now can be manually entered by supplying the start time, end time, and description for the timecard.

In addition, I modified how the forms are built so that both forms are defined inside the typescript file and you connect it to the html page using "[formGroup]" for each table and accessing the fields using formControlName="fieldNameDefinedInTypescriptFile"

There however still no check to make sure the start date is latter then the end date.  But it should be easily implemented by simply comparing the
two fields when you submit it.  Or more complicated, when before the user presses the submit, I however don't know how this would be don at the moment.