Manual Time card limit creation in place, but time difference/date processing currently not working right as intended
Manual time card limit now mostly implemented.  See below bug for what needs to be done.

Modified the seed.js file for the table TimeCard so it contains two new fields, timeCardCreation and isManualEntry so it can determine if the user has reached the 5 manual time cards for the day.  I have also modified the auto timecard creation inside the TS file so it now makes use of the new variables.  The field timeCardCreation is processed server side so the students can't game the system.

You have to drop the TimeCard table using the following SQL statement "DROP TABLE TimeCard;" and rebuild it for the new fields to be added.

There is also currently an issue with the logic for the manual entry fields, like if you have a time difference of 2 hours and 3 minutes, it will only include the 2 hours, as shown in the bug report.  Need to see why this is happening.