Time card entry form now restores to previous state
Now the forms that you use to make a time card for a project will restore itself to the last state you used it.
I.E. if you manually entered a time card and then left the project page and  came back to it, the manual form would be displayed instead of the auto one.