Date validation for the manual input is now partially implmented
Now the date fields for the manual input will check to make sure that the starting time is less then the ending time and they are not the time.
As of right now, it simply doesn't allow you to submit it, I.E. you press the button and nothing happens.  I still need to implement a way of displaying the conflicts to the user.  Possible something like what is described here https://blog.angular-university.io/angular-custom-validators/#:~:text=this%20function%20call.-,Displaying%20error%20messages%20for%20custom%20Validators,-In%20order%20to