Fixed bug with calculation of LNMidNight and TNmidnight variables
This bug made it so that after around 9 P.M. it would suddenly add an extra day to the variables.  This could make it so that users could enter as many time cards as they wanted after 9 P.M. until midnight.

The issue was that I was getting the UTC year, month, day from the variable currentDate while it was already in UTC.  This then would then take the difference between the current timezone and the UTC timezone and add it to the two variables.

The fix to this is to simply grab the local year, month, day from the variable because it is already in UTC.