Message now displayed when default admin account is created
Now when the Node server inserts the default admin account to the DB, the first person that goes to the login page will be greeted with an alert message saying that the default admin account has been generated.  Every other user that goes to the login page will not receive this message.

IMPORTANT:  You have to install a new package, node-localstorage, in  the Node server for this to work.  All you have to do is to run either the "npm i" command while inside the Node directory or simply run the "Install_Packages" PowerShell script file.