Implmented certificates that now will allow a secure connection to the servers
To make the computer consider these connections to the servers be a valid connection, there is one of two ways of doing this.

1.	Search for “manage user certificates” in your search bar.  This should open the certificate manager for the current user on the PC.  Next, navigate to the folder “Trusted Root Certification Authorities” > “Certificates” and right click on the “Certificates” folder and select the option “All Tasks” > “Import…”  Then simply follow the import wizard until it asks you to locate the file you wish to import.  Here you then browse to the certificates folder in both servers and select the .crt file.  Now you simply need to continue through the process and it will add the certificates to the PC.

2.	Navigate to the “certificates” folders for both the servers and double click on the .crt files to open them.  Next, select the “Install Certificate…” button and in the resulting window select the “Current User” radio button.  In the following window, select the “Place all certificates in the following store” radio button and then browse for the folder named “Trusted Root Certification Authorities”.  Then simply continue the process and it will add it to the PC.

Both of these methods will add these certificates to the certificate manager for the PC with the names of “localhost 4200” and “localhost 8080” for both the ‘Issued To” and the “Issued By” columns.  However, if you have your browser open, you have to close/restart it to have these changes go into effect.  A helpful command if you use Chrome is to type the following string into the URL and it will restart the browser without having anything closed “chrome://restart”.  Now when you connect to https://localhost:8080 or https://localhost:4200, the connection will say it is valid and secure.