Login/Register now uses the new hashing that resetting the password uses
This also makes it so that the password are hashed and form validation are done client side instead of server side.

This set of changes will make it so any existing users will be unable to log in.  So you have delete or drop every user from the Users table inside the DB.