GetUserTimeForProject now gets all timecards, not just those that are currently part of the group
Modified the GetUserTimesForProject route so that it will get all the timecards for the project regardless of if a user is part of the project.
I.E. if they created a time card for said project, they will always be shown in the team times regardless if they are in the project or not.