Editing a user profile now modifies the profile you are viewing instead of the current user's
This problem only affected the admins, but it is now fixed.
The changes inside of the edit-profile.componet.ts file seems like a lot, but the only thing that changes was the variable "currentUser" inside the consturctor and replacing it with what is in there now.  Everything else was just re-organized.