Modified some stuff dealing with evals
Re-orginzed and fixed some issues with the seed.js file.  I.E. Fixed a foreign key mis-match type and reorginzed it so that they it follows the pattern of tables that don't have foreign keys are generated first and then any other tables that use forien keys are then made.

Also, removed some unnessary code for the manage-evals component.