Logic for selecting projects to assign evals implmented
Now when you select a project from the "Projects" list, it will update some backend variables to keep track of what projects have been selected and what should be sent to the server.
As of right now, it simply prints out the information to the console.  That portion of the logic still needs to be implmented.