Assign Eval component now allows you to select a eval template
Additional changes:
Made it so that both at least one project and a eval template has to be selected before you can press the "Assign Evaluation" button.
Also made the logic needed to make the body of the HTTP put request to send the information to the server.  However, the API call still needs to be created to assign the eval to the project/project's users.