Cleaned up the code on the manage-evals compoent and added an additional controller for managing evals
Now inside the manage evals component, instead of having two option HTML components and hard coding the values for each one.  We now have it use ngFor to loop through every entry inside of the variable questionTypes and dynamicaly create the value and text for the combox.

I also made it so that the HTML and DB Question_Type table are more unified. I.E. instead of having Rating, 1-5 Rating, Text, and Text Response, we now simply have just 1-5 Rating and Text Response.