Assigning evals mostly implmented
The only thing that is not implmented is the correct error detection for inserting the data into the DB on lines 348-358.
Because currently, it will first get to the if conditions below the containing foreach loop and check the variable "errorsGenerated" before it even gets into the if/else statment for the db.run command.  Or at the very least, I think it doesn't work correctly.  I never tested it for when a insert is invalid.